requirements.txt
roboticstoolbox-python>=1.1.1
spatialmath-python>=0.10.0
numpy>=1.26.4
opencv-python>=4.11.0.86
matplotlib>=3.7.4
